+++ src/main/java/KUSITMS/WITHUS/domain/mailTemplate/dto/MailTemplateResponseDTO.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.global.infra.email.mailTemplate.dto;
+package KUSITMS.WITHUS.domain.mailTemplate.dto;
 
-import KUSITMS.WITHUS.global.infra.email.mailTemplate.entity.MailTemplate;
+import KUSITMS.WITHUS.domain.mailTemplate.entity.MailTemplate;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "메일 템플릿 응답 DTO")
