+++ src/main/java/KUSITMS/WITHUS/domain/mailTemplate/repository/MailTemplateJpaRepository.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.global.infra.email.mailTemplate.repository;
+package KUSITMS.WITHUS.domain.mailTemplate.repository;
 
-import KUSITMS.WITHUS.global.infra.email.mailTemplate.entity.MailTemplate;
+import KUSITMS.WITHUS.domain.mailTemplate.entity.MailTemplate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MailTemplateJpaRepository extends JpaRepository<MailTemplate, Long> {
