--- src/main/java/KUSITMS/WITHUS/global/infra/email/mailTemplate/service/MailTemplateService.java
@@ -1,11 +1,0 @@
-package KUSITMS.WITHUS.global.infra.email.mailTemplate.service;
-
-import KUSITMS.WITHUS.global.infra.email.mailTemplate.dto.MailTemplateRequestDTO;
-import KUSITMS.WITHUS.global.infra.email.mailTemplate.dto.MailTemplateResponseDTO;
-
-import java.util.List;
-
-public interface MailTemplateService {
-    List<MailTemplateResponseDTO.Summary> listAll();
-    MailTemplateResponseDTO.Detail create(MailTemplateRequestDTO.Create dto);
-}
